+++ components/hal/esp32s2/include/hal/aes_ll.h
@@ -49,7 +49,7 @@ static inline uint8_t aes_ll_write_key(const uint8_t *key, size_t key_word_len)
     volatile uint8_t key_in_hardware = 0;
     /* Memcpy to avoid potential unaligned access */
     uint32_t key_word;
-    for (int i = 0; i < key_word_len; i++) {
+    for (int i = 0; i < (int)key_word_len; i++) {
         memcpy(&key_word, key + 4 * i, 4);
         REG_WRITE(AES_KEY_BASE + i * 4,  key_word);
         key_in_hardware += 4;
